Move project files getter
