+++ Horologion/src/files.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+
+namespace files {
+std::string get_project_config();
+std::string get_project_log();
+} // namespace files
